[FIX] A session is created every time a page is refreshed
