+++ Gemfile
@@ -8,6 +8,7 @@ gem 'octokit', '~> 4.0'
 gem 'puma'
 gem 'redis'
 gem 'redis-rack'
+gem 'require_all'
 gem 'rest-client', '~> 1.6.3'
 gem 'sinatra'
 gem 'sinatra-contrib'
