+++ app/datasources/github.rb
@@ -1,0 +1,11 @@
+require './app/datasources/github/repositories'
+require './app/datasources/github/authentication'
+
+
+module Datasources
+  ##
+  # Github datasources namespace
+  #
+  module Github
+  end
+end
