+++ app/helpers/client.rb
@@ -1,0 +1,11 @@
+require 'octokit'
+
+module Helpers
+  ##
+  # Github client helper
+  module Client
+    def client
+      @client ||= Octokit::Client.new(:access_token => session[:access_token])
+    end
+  end
+end
