--- app/tentacles.rb
@@ -1,17 +1,0 @@
-require 'bundler/setup'
-require 'json'
-require 'rest_client'
-require 'sinatra'
-require 'sinatra/base'
-require 'tentacles/application'
-require 'tentacles/controllers'
-require 'tentacles/datasources'
-require 'tentacles/helpers'
-require 'tentacles/repositories'
-require 'tentacles/views'
-
-##
-# Tentacles app entrypoint
-#
-module Tentacles
-end
