--- app/tentacles/controllers.rb
@@ -1,11 +1,0 @@
-require 'tentacles/controllers/authentication'
-require 'tentacles/controllers/repositories'
-require 'tentacles/controllers/pulls'
-
-module Tentacles
-  ##
-  # Controllers Namespace
-  #
-  module Controllers
-  end
-end
