--- app/tentacles/datasources.rb
@@ -1,9 +1,0 @@
-require 'tentacles/datasources/github'
-
-module Tentacles
-  ##
-  # Datasources namespace
-  #
-  module Datasources
-  end
-end
