--- app/tentacles/datasources/github.rb
@@ -1,11 +1,0 @@
-require 'tentacles/datasources/github/repositories'
-require 'tentacles/datasources/github/authentication'
-
-module Tentacles
-  module Datasources
-    ##
-    # Github datasources namespace
-    module Github
-    end
-  end
-end
