--- app/tentacles/helpers.rb
@@ -1,11 +1,0 @@
-require 'tentacles/helpers/state'
-require 'tentacles/helpers/authentication'
-require 'tentacles/helpers/client'
-
-module Tentacles
-  ##
-  # Helpers Namespace
-  #
-  module Helpers
-  end
-end
