--- app/tentacles/helpers/client.rb
@@ -1,13 +1,0 @@
-require 'octokit'
-
-module Tentacles
-  module Helpers
-    ##
-    # Github client helper
-    module Client
-      def client
-        @client ||= Octokit::Client.new(:access_token => session[:access_token])
-      end
-    end
-  end
-end
