--- app/tentacles/repositories.rb
@@ -1,10 +1,0 @@
-require 'tentacles/repositories/repositories'
-require 'tentacles/repositories/user'
-
-module Tentacles
-  ##
-  # Repositories Namespace
-  #
-  module Repositories
-  end
-end
