--- app/tentacles/views.rb
@@ -1,7 +1,0 @@
-module Tentacles
-  ##
-  # Views Namespace
-  #
-  module Views
-  end
-end
