+++ config.ru
@@ -1,25 +1,19 @@
-require 'bundler'
 require 'dotenv'
-require 'rack'
-require 'sinatra'
-require 'sinatra/base'
+require 'require_all'
 require 'rack/session/redis'
 
-Dotenv.load
+ENV['SINATRA_ENV'] ||= 'development'
+
+require 'bundler/setup'
+Bundler.require(:default, ENV['SINATRA_ENV'])
+
+require_all 'app'
 
 use Rack::Session::Redis,
     :redis_server => "#{ENV.fetch('REDIS_URL')}/0/rack:session"
 
-Bundler.setup(:default)
-$LOAD_PATH.unshift("#{File.dirname(__FILE__)}/app")
+use Controllers::AuthenticationController
+use Controllers::RepositoriesController
+use Controllers::PullsController
 
-require 'tentacles'
-
-url_map = {
-  '/' => Tentacles::Application,
-  '/authentication' => Tentacles::Controllers::Authentication.new,
-  '/repositories' => Tentacles::Controllers::Repositories.new,
-  '/pulls' => Tentacles::Controllers::Pulls.new
-}
-
-run Rack::URLMap.new url_map
+run ApplicationController
